computer-vision/src/demo/requirements.txt
streamlit>=0.84.0
pandas>=1.1.0
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
plotly>=4.0.0
seaborn>=0.11.0
matplotlib>=3.3.0
ultralytics>=8.0.0
streamlit-drawable-canvas-fix>=0.9.4 